+++ android/src/main/java/com/artirigo/kontaktio/BeaconProximityManager.java
@@ -1,5 +1,8 @@
 package com.artirigo.kontaktio;
 
+import java.util.Collections;
+import java.util.UUID;
+
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReadableArray;
@@ -37,6 +40,13 @@ class BeaconProximityManager {
 
         // create manager
         proximityManager = ProximityManagerFactory.create(reactAppContext);
+
+        // this forces the new Anchor 2 beacons to include `uniqueId` with beacon ranging data
+        String proximityUuid = String.valueOf(KontaktSDK.DEFAULT_KONTAKT_BEACON_PROXIMITY_UUID);
+        proximityManager
+                .spaces()
+                .forceResolveRegions(Collections.singletonList(UUID.fromString(proximityUuid)))
+                .forceResolveNamespaces(Collections.singletonList(proximityUuid.refplace("-", "")));
 
         // Instantiate helpers
         beaconListeners = new BeaconListeners(reactAppContext);